+++ app/src/main/java/place/pic/data/remote/request/RequestToPlacceIdx.kt
@@ -1,5 +1,5 @@
 package place.pic.data.remote.request
 
 data class RequestToPlacceIdx(
-    val placeIdx: Int
+    val placeIdx: Long
 )
